Fix UB in NtUnicodeStrMut::try_from_u16()
Avoid transmuting an NtUnicodeStr into a NtUnicodeStrMut which is
assigned a *const instead of a *mut for its buffer field.
